+++ src/index.ts
@@ -2,4 +2,5 @@ export * from "./types.js";
 export * from "./continuity.js";
 export * from "./representation.js";
 export * from "./render-adapter.js";
+export * from "./solver.js";
 export * from "./worker.js";